requirements.txt
accelerate==1.1.1
diffusers==0.31.0
fastapi==0.115.5
huggingface-hub==0.26.2
python-dotenv==1.0.1
requests==2.32.3
scipy==1.14.1
tqdm==4.67.0
transformers==4.46.3
uvicorn==0.32.1
